azure-one/functions/requirements.txt
azure-functions>=1.18.0
azure-storage-blob>=12.19.0
azure-data-tables>=12.4.0
feedparser>=6.0.10
beautifulsoup4>=4.12.0
lxml>=5.1.0
requests>=2.31.0
pyyaml>=6.0.1
azure-identity>=1.15.0
azure-monitor-querymetrics>=1.0.0
